requirements.txt
streamlit
openai
pinecone
edge-tts
PyPDF2
python-docx
langchain-text-splitters
streamlit-mic-recorder
groq
python-dotenv